% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{autoWekaClassifier}
\alias{autoWekaClassifier}
\alias{print.RAutoWEKAClassifier}
\alias{summary.RAutoWEKAClassifier}
\alias{predict.RAutoWEKAClassifier}
\alias{model.frame.RAutoWEKAClassifier}
\alias{terms.frame.RAutoWEKAClassifier}
\title{Auto-WEKA Classifier}
\usage{
autoWekaClassifier(formula, data, subset, na.action, seed = 123,
  timeLimit = 15, memoryLimit = 1024, numberOfConfigs = 1,
  metric = c("errorRate", "areaAboveROC", "areaUnderROC", "avgCost",
  "correct", "correlationCoefficient", "falseNegativeRate",
  "falsePositiveRate", "fMeasure", "incorrect", "kappa", "kBInformation",
  "kBMeanInformation", "kBRelativeInformation", "meanAbsoluteError",
  "pctCorrect", "pctIncorrect", "precision", "relativeAbsoluteError",
  "rootMeanSquaredError", "rootRelativeSquaredError",
  "weightedAreaUnderROC", "weightedFalseNegativeRate",
  "weightedFalsePositiveRate", "weightedFMeasure", "weightedPrecision",
  "weightedRecall",      "weightedTrueNegativeRate",
  "weightedTruePositiveRate"), parallelRuns = 1,
  outputDebugInfo = FALSE, noCapabiltiesCheck = FALSE,
  decimalPlaces = 2, batchSize = 100)

\method{print}{RAutoWEKAClassifier}(x, ...)

\method{summary}{RAutoWEKAClassifier}(object, ...)

\method{predict}{RAutoWEKAClassifier}(object, newdata, type = c("class",
  "probability"), ...)

\method{model.frame}{RAutoWEKAClassifier}(formula, ...)

\method{terms}{frame.RAutoWEKAClassifier}(x, ...)
}
\arguments{
\item{formula}{a model formua.}

\item{data}{the data to build the classifier on.}

\item{subset}{an optional vector specifying a subset of observations to be used when building the classifier (see \code{\link[stats]{model.frame}}).}

\item{na.action}{a function that specifies how to deal with NAs (see \code{\link[stats]{model.frame}}).}

\item{seed}{the seed for the random number generator (defaults to 123).

Original parameter name: \code{-seed}.}

\item{timeLimit}{the (approximate) time limit for tuning in minutes (defaults to 15).

Original parameter name: \code{-timeLimit}.}

\item{memoryLimit}{the memory limit for runs in MiB (defaults to 1024).

Original parameter name: \code{-memLimit}.}

\item{numberOfConfigs}{the amount of best configurations to output (defaults to 1).

Original parameter name: \code{-nBestConfigs}.}

\item{metric}{the metric to optimise (defaults to \code{errorRate}). Available metrics:

\code{areaAboveROC}, \code{areaUnderROC}, \code{avgCost}, \code{correct}, \code{correlationCoefficient}, \code{errorRate}, \code{falseNegativeRate}, \code{falsePositiveRate}, \code{fMeasure}, \code{incorrect}, \code{kappa}, \code{kBInformation}, \code{kBMeanInformation}, \code{kBRelativeInformation}, \code{meanAbsoluteError}, \code{pctCorrect}, \code{pctIncorrect}, \code{precision}, \code{relativeAbsoluteError}, \code{rootMeanSquaredError}, \code{rootRelativeSquaredError}, \code{weightedAreaUnderROC}, \code{weightedFalseNegativeRate}, \code{weightedFalsePositiveRate}, \code{weightedFMeasure}, \code{weightedPrecision}, \code{weightedRecall}, \code{weightedTrueNegativeRate}, \code{weightedTruePositiveRate}.

Original parameter name: \code{-metric}.}

\item{parallelRuns}{the number of parallel runs (defaults to 1). Caution: this option is experimental.

Original parameter name: \code{-parallelRuns}.}

\item{outputDebugInfo}{logical indicating to run the classifier in debug mode which may output additional information (defaults to \code{FALSE}).

Original parameter name: \code{-output-debug-info}.}

\item{noCapabiltiesCheck}{logical indicating to not check the classifier capabilities before the classifier is built (defaults to \code{FALSE}). Use with caution.

Original parameter name: \code{-do-not-check-capabilities}.}

\item{decimalPlaces}{the number of decimal places for the output of numbers in the model (defaults to 2).

Original parameter name: \code{-num-decimal-places}.}

\item{batchSize}{the desired batch size for batch prediction (defaults to 100).

Original parameter name: \code{-batch-size}.}

\item{newdata}{the data to classify.}

\item{type}{the prediction type to output (defaults to \code{class}). Available types:
\describe{
  \item{\code{class}}{to get class memberships}
  \item{\code{probability}}{to get class membership probabilities}
}}
}
\value{
The Auto-WEKA classifier built.
}
\description{
This functions builds an Auto-WEKA classifier.
}
\examples{
\dontrun{
## Iris Example
data(iris)
# Build classifier with time limit of 5 minutes
classifier <- autoWekaClassifier(Species ~ ., iris, timeLimit = 5)
# Get summary
summary(classifier)
# Predict class memberships
predict(classifier, iris[1:10,-5])
# Predict class membership probabilities
predict(classifier, iris[1:10,-5], type = "probability")

## mlbench Exmaple
require("mlbench")
data("HouseVotes84", package = "mlbench")
# Build classifier with changed metric
classifier <- autoWekaClassifier(Class ~ ., HouseVotes84, metric = "precision")
predict(model, HouseVotes84[1:10, -1])
predict(model, HouseVotes84[1:10, -1], type = "probability")
}
}
\references{
C. Thornton, F. Hutter, H. H. Hoos, and K. Leyton-Brown (2013). Auto-WEKA: Combined selection and hyperparameter optimization of classification algorithms. In \emph{Proceedings of the 19th ACM SIGKDD international conference on Knowledge discovery and data mining} (pp. 847-855). ACM.

F. Hutter, L. Kotthoff, and J. Vanschoren (2019). Automatic Machine Learning: Methods, Systems, Challenges.

K. Hornik, C. Buchta, and A. Zeileis (2009). Open-source machine learning: R meets Weka. \emph{Computational Statistics}, \bold{24}/2, 225--232.
}
\author{
Implementation partially based on \code{\link[RWeka]{make_Weka_classifier}}.
}
