% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{buildAutoWekaClassifier}
\alias{buildAutoWekaClassifier}
\title{Build Auto-WEKA classifier}
\usage{
buildAutoWekaClassifier(trainingFile, testingFile, classIndex = "last",
  numberOfFolds = 10, noCrossValidation = FALSE,
  forceBatchTraining = FALSE, splitPercentage, preserveOrder = FALSE,
  randomNumberSeed = 1, costMatrixFile, toggle, modelInputFile,
  modelOutputFile, noStatistics = FALSE, onlyStatistics = FALSE,
  noPerClassStatistics = FALSE, informationTheoreticStatistics = FALSE,
  classificationOutputType, cumulativeMarginDistribution = FALSE,
  optionsXml, thresholdFile, thresholdLabel, noPredictions = FALSE,
  seed = 123, timeLimit = 15, memoryLimit = 1024,
  numberOfConfigs = 1, metric = "errorRate", parallelRuns = 1,
  outputDebugInfo = FALSE, noCapabiltiesCheck = FALSE,
  decimalPlaces = 2, batchSize = 100)
}
\arguments{
\item{trainingFile}{the name of the file containing the training data.

Original parameter name: \code{-t}.}

\item{testingFile}{the name of the file containing the test data. If missing, a cross-validation will be performed on the training data.

Original parameter name: \code{-T}.}

\item{classIndex}{the index of the class attribute (defaults to last). Possible values are \code{first}, \code{last} or the actual index.

Original parameter name: \code{-c}.}

\item{numberOfFolds}{the number of folds for cross-validation (defaults to 10).

Original parameter name: \code{-x}.}

\item{noCrossValidation}{logical indicating if cross validation should be pervormed (defualts to \code{FALSE}).

Original parameter name: \code{-no-cv}.}

\item{forceBatchTraining}{logical indicating to train classifier in batch mode instead of incrementally (defualts to \code{FALSE}).

Original parameter name: \code{-force-batch-training}.}

\item{splitPercentage}{the percentage for splitting into train/test set (e.g. 66).

Original parameter name: \code{-split-percentage}.}

\item{preserveOrder}{logical indicating to preserve the order in the percentage split (defaults to \code{FALSE}).

Original parameter name: \code{-preserve-order}.}

\item{randomNumberSeed}{the random number seed for cross-validation or percentage split (defaults to 1).

Original parameter name: \code{-s}.}

\item{costMatrixFile}{the name of the file containing the cost matrix.

Original parameter name: \code{-m}.}

\item{toggle}{a comma separated list of metric names to toggle in the output. All metrics are output by default with the exception of 'Coverage' and 'Region size'. Available metrics:

Correct, Incorrect, Kappa, Total cost, Average cost, KB relative, KB information, Correlation, Complexity 0, Complexity scheme, Complexity improvement, MAE, RMSE, RAE, RRSE, Coverage, Region size, TP rate, FP rate, Precision, Recall, F-measure, MCC, ROC area, PRC area

Original parameter name: \code{-toggle}.}

\item{modelInputFile}{the name of the model input file. In case the filename ends with \code{.xml}, a PMML file is loaded or, if that fails, options are loaded from the XML file.

Original parameter name: \code{-l}.}

\item{modelOutputFile}{the name of the model output file. In case the filename ends with \code{.xml}, only the options are saved to the XML file, not the model.

Original parameter name: \code{-d}.}

\item{noStatistics}{logical indicating not to output statistics for training data (defaults to \code{FALSE}).

Original parameter name: \code{-v}.}

\item{onlyStatistics}{logical indicating to output statistics only, not the classifier (defaults to \code{FALSE}).

Original parameter name: \code{-o}.}

\item{noPerClassStatistics}{logical indicating to not output statistics for each class (defaults to \code{FALSE}).

Original parameter name: \code{-do-not-output-per-class-statistics}.}

\item{informationTheoreticStatistics}{logical indicating to output information-theoretic statistics (defaults to \code{FALSE}).

Original parameter name: \code{-k}.}

\item{classificationOutputType}{the type of output to generate for the classification. Available types:

CSV, CSVTab, HTML, PlainText, XML, Null

Original parameter name: \code{-classifications}.}

\item{cumulativeMarginDistribution}{logical indicating to only output the cumulative margin distribution.

Original parameter name: \code{-r}.}

\item{optionsXml}{the name of a XML file or a XML string contianing the options.

Original parameter name: \code{-xml}.}

\item{thresholdFile}{the name of the file to save the threshold data to. The format is determined by the file extension (e.g. \code{.arff} for ARFF format or \code{.csv} for CSV).

Original parameter name: \code{-threshold-file}.}

\item{thresholdLabel}{the class label to determine the threshold data for (defaults to the first label).

Original parameter name: \code{-threshold-label}.}

\item{noPredictions}{logical indicating to not collect predictions in order to conserve memory (defaults to \code{FALSE}).

Original parameter name: \code{-no-predictions}.}

\item{seed}{the seed for the random number generator (defaults to 123).

Original parameter name: \code{-seed}.}

\item{timeLimit}{the (approximate) time limit for tuning in minutes (defaults to 15).

Original parameter name: \code{-timeLimit}.}

\item{memoryLimit}{the memory limit for runs in MiB (defaults to 1024).

Original parameter name: \code{-memLimit}.}

\item{numberOfConfigs}{the amount of best configurations to output (defaults to 1).

Original parameter name: \code{-nBestConfigs}.}

\item{metric}{the metric to optimise (defaults to \code{errorRate}). Available metrics:


areaAboveROC, areaUnderROC, avgCost, correct, correlationCoefficient, errorRate, falseNegativeRate, falsePositiveRate, fMeasure, incorrect, kappa, kBInformation, kBMeanInformation, kBRelativeInformation, meanAbsoluteError, pctCorrect, pctIncorrect, precision, relativeAbsoluteError, rootMeanSquaredError, rootRelativeSquaredError, weightedAreaUnderROC, weightedFalseNegativeRate, weightedFalsePositiveRate, weightedFMeasure, weightedPrecision, weightedRecall, weightedTrueNegativeRate, weightedTruePositiveRate

Original parameter name: \code{-metric}.}

\item{parallelRuns}{the number of parallel runs (defaults to 1). Caution: this option is experimental.

Original parameter name: \code{-parallelRuns}.}

\item{outputDebugInfo}{logical indicating to run the classifier in debug mode which may output additional information (defaults to \code{FALSE}).

Original parameter name: \code{-output-debug-info}.}

\item{noCapabiltiesCheck}{logical indicating to not check the classifier capabilities before the classifier is built (defaults to \code{FALSE}). Use with caution.

Original parameter name: \code{-do-not-check-capabilities}.}

\item{decimalPlaces}{the number of decimal places for the output of numbers in the model (defaults to 2).

Original parameter name: \code{-num-decimal-places}.}

\item{batchSize}{the desired batch size for batch prediction (defaults to 100).

Original parameter name: \code{-batch-size}.}
}
\value{
an Auto-WEKA classifier
}
\description{
This functions builds an Auto-WEKA classifier.
}
\examples{
\dontrun{
buildAutoWekaClassifier("http://storm.cis.fordham.edu/~gweiss/data-mining/weka-data/iris.arff", timeLimit = 1, noCrossValidation = TRUE)
}
}
