% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{autoWekaClassifier}
\alias{autoWekaClassifier}
\alias{print.RAutoWEKAClassifier}
\alias{summary.RAutoWEKAClassifier}
\alias{predict.RAutoWEKAClassifier}
\alias{model.frame.RAutoWEKAClassifier}
\alias{terms.RAutoWEKAClassifier}
\title{Auto-WEKA Classifier}
\usage{
autoWekaClassifier(formula, data, subset, na.action, seed = 123,
  timeLimit = 15, memoryLimit = 1024, numberOfConfigs = 1,
  metric = c("errorRate", "areaAboveROC", "areaUnderROC", "avgCost",
  "correct", "correlationCoefficient", "falseNegativeRate",
  "falsePositiveRate", "fMeasure", "incorrect", "kappa", "kBInformation",
  "kBMeanInformation", "kBRelativeInformation", "meanAbsoluteError",
  "pctCorrect", "pctIncorrect", "precision", "relativeAbsoluteError",
  "rootMeanSquaredError", "rootRelativeSquaredError",
  "weightedAreaUnderROC", "weightedFalseNegativeRate",
  "weightedFalsePositiveRate", "weightedFMeasure", "weightedPrecision",
  "weightedRecall",      "weightedTrueNegativeRate",
  "weightedTruePositiveRate"), parallelRuns = 1,
  outputDebugInfo = FALSE)

\method{print}{RAutoWEKAClassifier}(x, ...)

\method{summary}{RAutoWEKAClassifier}(object, ...)

\method{predict}{RAutoWEKAClassifier}(object, newdata, type = c("class",
  "probability"), ...)

\method{model.frame}{RAutoWEKAClassifier}(formula, ...)

\method{terms}{RAutoWEKAClassifier}(x, ...)
}
\arguments{
\item{formula}{a model formua.}

\item{data}{the data to build the classifier on.}

\item{subset}{an optional vector specifying a subset of observations to be used when building the classifier (see \code{\link[stats]{model.frame}}).}

\item{na.action}{a function that specifies how to deal with NAs (see \code{\link[stats]{model.frame}}).}

\item{seed}{the seed for the random number generator (defaults to 123).

Original parameter name: \code{-seed}.}

\item{timeLimit}{the (approximate) time limit for tuning in minutes (defaults to 15).

Original parameter name: \code{-timeLimit}.}

\item{memoryLimit}{the memory limit for runs in MiB (defaults to 1024).

Original parameter name: \code{-memLimit}.}

\item{numberOfConfigs}{the amount of best configurations to output (defaults to 1).

Original parameter name: \code{-nBestConfigs}.}

\item{metric}{the metric to optimise (defaults to \code{errorRate}). Available metrics:

\code{areaAboveROC}, \code{areaUnderROC}, \code{avgCost}, \code{correct}, \code{correlationCoefficient}, \code{errorRate}, \code{falseNegativeRate}, \code{falsePositiveRate}, \code{fMeasure}, \code{incorrect}, \code{kappa}, \code{kBInformation}, \code{kBMeanInformation}, \code{kBRelativeInformation}, \code{meanAbsoluteError}, \code{pctCorrect}, \code{pctIncorrect}, \code{precision}, \code{relativeAbsoluteError}, \code{rootMeanSquaredError}, \code{rootRelativeSquaredError}, \code{weightedAreaUnderROC}, \code{weightedFalseNegativeRate}, \code{weightedFalsePositiveRate}, \code{weightedFMeasure}, \code{weightedPrecision}, \code{weightedRecall}, \code{weightedTrueNegativeRate}, \code{weightedTruePositiveRate}.

Original parameter name: \code{-metric}.}

\item{parallelRuns}{the number of parallel runs (defaults to 1). Caution: this option is experimental.

Original parameter name: \code{-parallelRuns}.}

\item{outputDebugInfo}{logical indicating to run the classifier in debug mode which may output additional information (defaults to \code{FALSE}).}

\item{newdata}{the data to classify.}

\item{type}{the prediction type to output (defaults to \code{class}). Available types:
\describe{
  \item{\code{class}}{to get class memberships}
  \item{\code{probability}}{to get class membership probabilities}
}}
}
\value{
The Auto-WEKA classifier built. The result is represented as a list of the following elements (if applicable):
\describe{
  \item{\code{call}}{the call that has been made on the function.}
  \item{\code{mf}}{the model frame.}
  \item{\code{terms}}{the terms of the given forumula.}
  \item{\code{instances}}{a S4 object holding the input data as WEKA instances.}
  \item{\code{awc}}{a S4 object representing the classifier built.}
  \item{\code{cor}}{the correlation coefficient.}
  \item{\code{mae}}{the mean absolute error.}
  \item{\code{rmse}}{the root mean square error.}
  \item{\code{rae}}{the relative absolute error.}
  \item{\code{rrse}}{the root relative rquared Error.}
  \item{\code{numInstances}}{the number of instances.}
  \item{\code{classPredicitions}}{the class memberships for the given test instances.}
  \item{\code{probabilityPredictions}}{the class membership probabilities for given test instances.}
}
In addition, the chosen metric is contained in the result list as well.
}
\description{
This functions builds an Auto-WEKA classifier. Note that Auto-WEKA also supports regression algorithms, although the function name would not suggest that. It has been decided to stick to Auto-WEKA's naming as it is in line with WEKA's terminology (which also only uses the term classifier due to historic reasons).
}
\examples{
\dontrun{
## Iris Example
data(iris)
# Build classifier with time limit of 5 minutes
classifier <- autoWekaClassifier(Species ~ ., iris, timeLimit = 5)
# Get summary
summary(classifier)
# Predict class memberships
predict(classifier, iris[1:10,-5])
# Predict class membership probabilities
predict(classifier, iris[1:10,-5], type = "probability")

## mlbench Exmaple
require("mlbench")
data("HouseVotes84", package = "mlbench")
# Build classifier with changed metric
classifier <- autoWekaClassifier(Class ~ ., HouseVotes84, metric = "precision")
predict(classifier, HouseVotes84[1:10, -1])
predict(classifier, HouseVotes84[1:10, -1], type = "probability")
}
}
\references{
C. Thornton, F. Hutter, H. H. Hoos, and K. Leyton-Brown (2013). Auto-WEKA: Combined selection and hyperparameter optimization of classification algorithms. In \emph{Proceedings of the 19th ACM SIGKDD international conference on Knowledge discovery and data mining} (pp. 847-855). ACM.

F. Hutter, L. Kotthoff, and J. Vanschoren (2019). Automatic Machine Learning: Methods, Systems, Challenges.

K. Hornik, C. Buchta, and A. Zeileis (2009). Open-source machine learning: R meets Weka. \emph{Computational Statistics}, \bold{24}/2, 225--232.

Eibe Frank, Mark A. Hall, and Ian H. Witten (2016). The WEKA Workbench. Online Appendix for "Data Mining: Practical Machine Learning Tools and Techniques", Morgan Kaufmann, Fourth Edition, 2016.
}
\author{
Implementation partially based on the function \code{\link[RWeka]{make_Weka_classifier}} contained in the \href{https://cran.r-project.org/package=RWeka}{RWeka} package.
}
