% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{buildAutoWekaClassifier}
\alias{buildAutoWekaClassifier}
\alias{print.RAutoWEKAClassifier}
\alias{summary.RAutoWEKAClassifier}
\alias{predict.RAutoWEKAClassifier}
\alias{model.frame.RAutoWEKAClassifier}
\alias{terms.frame.RAutoWEKAClassifier}
\title{Build Auto-WEKA classifier}
\usage{
buildAutoWekaClassifier(formula, data, subset, na.action, seed = 123,
  timeLimit = 15, memoryLimit = 1024, numberOfConfigs = 1,
  metric = "errorRate", parallelRuns = 1, outputDebugInfo = FALSE,
  noCapabiltiesCheck = FALSE, decimalPlaces = 2, batchSize = 100)

\method{print}{RAutoWEKAClassifier}(x, ...)

\method{summary}{RAutoWEKAClassifier}(object, ...)

\method{predict}{RAutoWEKAClassifier}(object, newdata, type = c("class",
  "probability"), ...)

\method{model.frame}{RAutoWEKAClassifier}(formula, ...)

\method{terms}{frame.RAutoWEKAClassifier}(x, ...)
}
\arguments{
\item{formula}{a model formua.}

\item{data}{the data to build the classifier on.}

\item{subset}{an optional vector specifying a subset of observations to be used when building the classifier (see \code{\link[stats]{model.frame}}).}

\item{na.action}{a function that specifies how to deal with NAs (see \code{\link[stats]{model.frame}}).}

\item{seed}{the seed for the random number generator (defaults to 123).

Original parameter name: \code{-seed}.}

\item{timeLimit}{the (approximate) time limit for tuning in minutes (defaults to 15).

Original parameter name: \code{-timeLimit}.}

\item{memoryLimit}{the memory limit for runs in MiB (defaults to 1024).

Original parameter name: \code{-memLimit}.}

\item{numberOfConfigs}{the amount of best configurations to output (defaults to 1).

Original parameter name: \code{-nBestConfigs}.}

\item{metric}{the metric to optimise (defaults to \code{errorRate}). Available metrics:

areaAboveROC, areaUnderROC, avgCost, correct, correlationCoefficient, errorRate, falseNegativeRate, falsePositiveRate, fMeasure, incorrect, kappa, kBInformation, kBMeanInformation, kBRelativeInformation, meanAbsoluteError, pctCorrect, pctIncorrect, precision, relativeAbsoluteError, rootMeanSquaredError, rootRelativeSquaredError, weightedAreaUnderROC, weightedFalseNegativeRate, weightedFalsePositiveRate, weightedFMeasure, weightedPrecision, weightedRecall, weightedTrueNegativeRate, weightedTruePositiveRate

Original parameter name: \code{-metric}.}

\item{parallelRuns}{the number of parallel runs (defaults to 1). Caution: this option is experimental.

Original parameter name: \code{-parallelRuns}.}

\item{outputDebugInfo}{logical indicating to run the classifier in debug mode which may output additional information (defaults to \code{FALSE}).

Original parameter name: \code{-output-debug-info}.}

\item{noCapabiltiesCheck}{logical indicating to not check the classifier capabilities before the classifier is built (defaults to \code{FALSE}). Use with caution.

Original parameter name: \code{-do-not-check-capabilities}.}

\item{decimalPlaces}{the number of decimal places for the output of numbers in the model (defaults to 2).

Original parameter name: \code{-num-decimal-places}.}

\item{batchSize}{the desired batch size for batch prediction (defaults to 100).

Original parameter name: \code{-batch-size}.}
}
\value{
an Auto-WEKA classifier.
}
\description{
This functions builds an Auto-WEKA classifier.
}
\examples{
classifier = buildAutoWekaClassifier(Species ~ ., iris, timeLimit = 1)
predict(classifier, iris[1:10,-5])

}
\seealso{
\code{\link[RWeka]{make_Weka_classifier}}
}
